Created proper services to include the ownership of member and registration.
